+++ shuup_tests/admin/test_shipment_creator.py
@@ -6,24 +6,25 @@
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
 
-import pytest
 import re
 
+import pytest
 from bs4 import BeautifulSoup
 from django import forms
 
+from shuup.admin.form_modifier import FormModifier
 from shuup.admin.modules.orders.views.shipment import (
-    FORM_MODIFIER_PROVIDER_KEY, OrderCreateShipmentView, ShipmentFormModifier
+    OrderCreateShipmentView, ShipmentForm
 )
 from shuup.apps.provides import override_provides
 from shuup.core.excs import NoShippingAddressException
-
 from shuup.core.models import Order
 from shuup.testing.factories import (
     create_order_with_product, create_product, get_default_shop,
     get_default_supplier
 )
 from shuup.testing.utils import apply_request_middleware
+
 
 @pytest.mark.django_db
 def test_shipment_creating_view_get(rf, admin_user):
@@ -79,7 +80,7 @@ def test_extending_shipment_with_extra_fields(rf, admin_user):
         product, supplier, quantity=quantity, taxless_base_unit_price=1, shop=shop)
 
     extend_form_class = "shuup_tests.admin.test_shipment_creator.ShipmentFormModifierTest"
-    with override_provides(FORM_MODIFIER_PROVIDER_KEY, [extend_form_class]):
+    with override_provides(ShipmentForm.form_modifier_provide_key, [extend_form_class]):
         request = apply_request_middleware(rf.get("/"), user=admin_user)
         view = OrderCreateShipmentView.as_view()
         response = view(request, pk=order.pk).render()
@@ -102,7 +103,7 @@ def test_extending_shipment_clean_hook(rf, admin_user):
         product, supplier, quantity=quantity, taxless_base_unit_price=1, shop=shop)
 
     extend_form_class = "shuup_tests.admin.test_shipment_creator.ShipmentFormModifierTest"
-    with override_provides(FORM_MODIFIER_PROVIDER_KEY, [extend_form_class]):
+    with override_provides(ShipmentForm.form_modifier_provide_key, [extend_form_class]):
         data = {
             "q_%s" % product.pk: 1,
             "supplier": supplier.pk,
@@ -126,7 +127,7 @@ def test_extending_shipment_form_valid_hook(rf, admin_user):
         product, supplier, quantity=quantity, taxless_base_unit_price=1, shop=shop)
 
     extend_form_class = "shuup_tests.admin.test_shipment_creator.ShipmentFormModifierTest"
-    with override_provides(FORM_MODIFIER_PROVIDER_KEY, [extend_form_class]):
+    with override_provides(ShipmentForm.form_modifier_provide_key, [extend_form_class]):
         phone_number = "+358911"
         data = {
             "q_%s" % product.pk: 1,
@@ -149,7 +150,7 @@ def test_extending_shipment_form_valid_hook(rf, admin_user):
         assert shipment.products.first().product_id == product.id
 
 
-class ShipmentFormModifierTest(ShipmentFormModifier):
+class ShipmentFormModifierTest(FormModifier):
     def get_extra_fields(self, order):
         return [("phone", forms.CharField(label="Phone", max_length=64, required=False))]
 
@@ -174,13 +175,13 @@ def test_shipment_creating_with_no_shipping_address(rf, admin_user):
     supplier = get_default_supplier()
     product = create_product(sku="test-sku", shop=shop, supplier=supplier, default_price=3.33)
     order = create_order_with_product(product, supplier, quantity=1, taxless_base_unit_price=1, shop=shop)
-    
-    # remove shipping address 
+
+    # remove shipping address
     order.shipping_address = None
     order.save()
 
     with pytest.raises(NoShippingAddressException):
         order.create_shipment_of_all_products()
-    
+
     # order should not have any shipments since it should have thrown an exception
     assert order.shipments.count() == 0