+++ doc/changelog.rst
@@ -20,6 +20,9 @@ Localization
 Admin
 ~~~~~
 
+- Make form modifiers reusable. Users of `ShipmentFormModifier`
+  should update any references to implement the
+  `shuup.admin.form_modifier.FormModifier` interface instead
 - Add mass actions to products list
 - Add mass actions to orders list
 - Add mass actions to contacts list